requirements.txt
python-ldap
python-nss
